Remove trace.{c,h}
The only thing left in there was `ddtrace_tracer_is_limited`, so moved
it into ddtrace.{c,h}.
Also did a tiny refactor on it to ensure that the addition does not
overflow (widen then add, not add then widen).
