+++ package.xml
@@ -75,8 +75,6 @@
                     <file name="signals.h" role="src" />
                     <file name="span.c" role="src" />
                     <file name="span.h" role="src" />
-                    <file name="trace.c" role="src" />
-                    <file name="trace.h" role="src" />
                     <file name="version.h" role="src" />
                     <file name="compatibility.h" role="src" />
                     <dir name="mpack">
